REFACTOR: Change Landing Page From Inventories to Login
In this commit, the landing page of this web application is changed From
the inventories page to the login page.

Only authenticated user is permitted to view the inventories page.

Before the backend communication is established, in order to navigate to
the inventories page, the username is "boss", and password can be anything